src/tic-tac-toe/predict-classification/requirements.txt
pip==23.1.2
setuptools==67.7.2
wheel==0.40.0
# Tensorflow and other package replace the numpy package for <1.24, for this reason I commented numpy.
# numpy==1.24.2
# numpy==1.23.5
pandas==2.0.1
matplotlib==3.7.1
scikit-learn==1.2.2
# tensorflow==2.12.0
tensorflow-cpu==2.12.0
# intel-tensorflow==2.12.0
keras==2.12.0
keras-tuner==1.3.5
